Updating Load Next Seed logic
This logic update allows the next seed function to handle ROM names with leading zeros (e.g. "Rom001", "Rom002"). It also handles the weird cause of having spaces and/or underscores in the Rom file name, which bizhawk makes difficult for no reason. There is a minor amount of error checking done here.