Refactored Theme-Color References
Huge overhaul for most everything that references a theme color. Now instead of storing the color directly in the object or the drawing, it references the color using an index. This allows for faster and simpler theme updates.

Also cleaned up most of the Theme code and added a safer Restore Defaults button.