Adding nature symbols and final theme
Adding color independent ways to know which nature is positive or negative. This is important given the uncertainty of end-users choosing theme colors that may not make this clear. As a result of this addition, I also cleaned up drawing the stats box a good bit.

Added another theme preset for six total. Might add more later.