Color picker doesn't cause theme to save
Fixing an issue where if the only change you make to your theme is through the color picker, then it doesn't properly store this info in the Settings object to be saved. This ensures any options/theme changes will ALL get written if the 'updated' flag on either is present.