Fix Catch Trace and Catch Evo bugs
This fixes an issue where catching a Pokemon with Trace records its ability as the "traced ability" and not actually the ability Trace.

Unrelated, this fixes an issue where catching a pokemon and then immediately evolving it won't remove it's old ability, because the personality values still matched.

Finally, doing some small renaming for frame updates. And saving tracker data after ever battle.