Tracker data now saves to a file
This is a first step in separating out the tracker data from save-states. There are a couple of reasons for this, as well as potential future improvements on allowing the end-user to better manage tracked data between games.

The primary reason for separate tracked data is some users do not or can not use save-states. As such, they fear losing their Tracker data and are forced to keep their computer/game running.

Secondly, some users run into an issue where Bizhawk causes a sound bug when from loading game data from a save state (unrelated to Tracker). They can use in-game saves to bypass this bug, but have no way to get their Tracker data back.