Fixing Focus Punch track on chargeup
Focus Punch is a weird move in that it doesn't deplete PP when it's charging up, but only if the pokemon gets a turn to execute the move (or flinch). If the pokemon dies, then punch is never tracked.

This fix checks the battle messages to see if the charge up animation occurs, and if so it tracks the move then and there. Confirmed to work in FireRed 1.0, unsure if it works in other versions.