Auto Badge Tracking support
Thanks to some research from thisisatest, we can now automatically check for when badges are obtained and auto update the tracker with this information. This feature works for RSE and FRLG.

Also disabling the focus punch check, as that has conditions where it fails and reveals too much info to the user.