Refactoring globals, Settings, Screens
Further organizing code such that it is more readable and intuitive to work with. Changes to when/how data is saved to the Settings file. Also non-functional changes to redraw and screen view.

Moving comments out of the Settings.ini file, as this shouldn't ever need to be opened by the user. Will make into a wiki later