TDAT files now display multi-lined
TDAT files were simply just a single line of text. Made a small change to the way it's written such that its split into multiple lines in the file, making it a tad easier to read (for humans). Non-functional change on the Tracker being able to read/write old TDAT files.