Removing Settings.ini to allow upgrade-safe
Settings.ini file is no longer required for first downloads of the Tracker. It will load with built-in defaults. If any changes are made or themes applied, the Settings file will still be created.

Thus, when a user inevitably upgrades to a new release, the Settings.ini file won't be part of that upgrade and won't overwrite their existing settings. New settings will still work and load as defaults.