Added support for Option toggle
Added an option toggle to Gameplay Options to enable or disable the revealing of randomized data. This includes all pokemon types, move types, move powers, move pps, and move accuracies. Each area of the Tracker that reveals this information will first check the Option toggle and verify its safe to show to the player.