Fixing old save data interfering with pregame
This is a quick-fix that verifies the player is actually playing the game before revealing their Pokemon team. This is because Pokemon team data is still read in for some reason during the game loading screen, before a save file  or new game is even selected.

Right now this checks based on mapId, which I'm assuming is 0 outside of game, and any other positive value while playing the game.