Adding lower box text color into theme
Theme strings now include a "Lower box text" color. If it's not present in the imported theme, then the "Default text" color hex code is duplicated and used.

This is required for an incoming update to the NDS tracker and the new theme strings that will be created as a result.