New Feature: Full randomized ball picker feature
The random ball picker only appears if the player doesn't have any Pokémon and when the player's character is in "the lab".  When it is first shown, it will display a randomly chosen ball. The user can click the re-roll dice button to re-randomize the selection.

This can be disabled through the Tracker settings "Show random ball picker".