Last seen level of encountered Pokemon now tracked
This adds support for tracking and displaying the last level seen for an encountered Pokemon. When encountering a Pokemon for the first time, a blank line is shown. Future encounters will show the level of the previously encountered Pokemon. This helps match move data tracked from previous encounters with the newly encountered, likely higher level Pokemon.

To make space for displaying this feature, the old "HP: ?/?" useless text has been removed when viewing the enemy Pokemon.