Updater: removing "use old download"
When an update failed to apply changes to files, it would keep the downloaded version on hand. When it ran again, it'd attempt to use those files. This works fine except for cases where an update fails a long time ago and those fails are never removed or refreshed to latest release.