Custom Code: New interface-like implementation
Splitting off the custom code file into two pieces: the core code file (which is now part of the tracker project), and the settings code file. The latter is part of the add-on that is used by the developer to enable custom code and supply a list of code files.