Extensions: re-enabling now loads new code
Prior to this, disabling and enabling an extension did NOT load the code file again (no need if it was the same). However, common practice is to update extension files then turn them off and on again to get the updates, thus manually enabling an extension should load the lua code file again.