Add specific logic for >1 evolotion. Old pattern caught it.
The old lua pattern managed to handle more than one evolution for a pokemon, but lua patterns are farily limited so instead of trying to make the pokemonName pattern more complex, I added logic in the code to explicitly handle the case where there are is than one evolution.
It handles the case of 2 evolutions e.g "Pikachu and Raichu" and the case of more than 2 evolutions e.g. "Pikachu, Raichu, Pichu, and Raiku".