LogViewer: moves show stab/categories
Adding in some extra information to the pokemon zoom view in the log viewer. This shows which moves belong to which category (physical/special/none), as well as colors moves that are stab for the viewed pokemon, just like on the tracker battle screen.

Due to this change, the way that TM Moves look is a bit different, and grouped to show gym tms first, then other learnable tms.

Also fixed an annoying issue where log viewer's right info screen kept redrawing gameover screen while in the Hall of Fame.