Fix LogViewer opening non-related logs
For cases where you load a rom that doesnt match the name of the rom defined in the quickload settings, perhaps playing Survival rom but the old quickload is still on Kaizo. This means the log autodetect wont load up the Kaizo rom (from quickload) but instead prompt user for a log file.