Remove "Remind me" feature from Update Screen
The "Remind me later" feature from before was carried over to the UpdateScreen reworks, but there is no clear and intuitive way for the user to know they've chosen to be reminded (clicking Back instead of Ignore). As such, I've temporarily disabled this feature, leaving in the code as commented out if needed to be used again.