Network send allowed events, and info w/ responses
When starting up, the network should inform the other connection point of the allowed/enabled events (commands and rewards). In this way, the client will only have to send these events back over to the Tracker, instead of ALL events.

Responses now allow for AdditionalInfo, such as sending ids for reward fulfillment and cancellation.