Click empty ability area now opens Notes
Previously, this was a bit unintuitive to click on the empty Ability section in order to set a tracked ability for a Pokémon. In order to do this, you would need to click the Notes area on the bottom carousel box. Now, you can simply click the Ability area to get the same popup Notes box.