Info Screen mon levels/exp reflects any battler
Previously this only checked against the pokemon being viewed on the main tracker screen. Toggling the view would make the levels/exp change, which felt weird.

Now it checks if the info screen pokemon is any of the battlers, and if so, users that battler's level for the moves learned and exp yield.