Code to track initial moveset of Pokemon used
This solves an outstanding issue where the initial 4 moves of a player's lab/pivot Pokemon are never tracked, since the conditions for tracking a move were "used in battle" or "learned while leveling up".

This remedies that issue by doing a 1-time tracking of the current moveset when battling a trainer (player decided to run this Pokemon for real). It also only tracks moves within its initial moveset if that Pokemon could have naturally learned that move by level up.