Override moves shown to use known moves if able
In some cases, while battling and viewing an enemy pokemon, that pokemon will use all four of its moves. When it does, show those four moves on the tracker screen instead of the loose ordering of otherwise "tracked moves" it used to show. This fixes the bug where 1 of the 4 moves wouldnt be shown even if all were known during that single battle.