Add profile gameover condition; auto select
Add the ability to choose how the game is considered over for different profiles with different rulesets. This is a redundant setting with the existing Gameplay options Game Over setting.

This is necessary to show at the profile creation/edit level, as the Game Over setting should be autodetected when setting up a new profile, such that a user wants to play Standard Ironmon