Finally add rom mismatch check
This adds a rom hash check when a ROM file is opened that doesn't match the previously played rom (based on stored rom hash in Tracker notes). When this occurs, a warning prompt is issued to the user requesting a mandator action to take: either fix it, or dismiss and deal with the consequences.

This warning is required for use with GachaMon, as profiles may change and wrong roms get loaded. If so, the recent mons remembered for the last played game shouldn't just get deleted.

Also add Dive to banned moves for GachaMon ratings (non-Standard rulesets)