Adjust rom mismatch check; ignore if same name
In the rare case where the two roms have the same exact name, the "fix it for me" almost never will work. if they're the same name, this likely means their original rom file was replaced and can't be recovered/restored.

as such, ignore the mismatch check if the filenames are known and different